+++ rage/src/bin/rage-keygen/main.rs
@@ -1,7 +1,7 @@
 #![forbid(unsafe_code)]
 
 use age::{cli_common::file_io, secrecy::ExposeSecret};
-use gumdrop::Options;
+use clap::{builder::Styles, ArgAction, Parser};
 use i18n_embed::{
     fluent::{fluent_language_loader, FluentLanguageLoader},
     DesktopLanguageRequester,
@@ -31,15 +31,34 @@ macro_rules! fl {
     }};
 }
 
-#[derive(Debug, Options)]
+#[derive(Debug, Parser)]
+#[command(display_name = "rage-keygen")]
+#[command(name = "rage-keygen")]
+#[command(version)]
+#[command(help_template = format!("\
+{{before-help}}{{about-with-newline}}
+{}{}:{} {{usage}}
+
+{{all-args}}{{after-help}}\
+    ",
+    Styles::default().get_usage().render(),
+    fl!("usage-header"),
+    Styles::default().get_usage().render_reset()))]
+#[command(next_help_heading = fl!("flags-header"))]
+#[command(disable_help_flag(true))]
+#[command(disable_version_flag(true))]
 struct AgeOptions {
-    #[options(help = "Print this help message and exit.")]
-    help: bool,
+    #[arg(action = ArgAction::Help, short, long)]
+    #[arg(help = fl!("help-flag-help"))]
+    help: Option<bool>,
 
-    #[options(help = "Print version info and exit.", short = "V")]
-    version: bool,
+    #[arg(action = ArgAction::Version, short = 'V', long)]
+    #[arg(help = fl!("help-flag-version"))]
+    version: Option<bool>,
 
-    #[options(help = "Write the result to the file at path OUTPUT. Defaults to standard output.")]
+    #[arg(short, long)]
+    #[arg(value_name = fl!("output"))]
+    #[arg(help = fl!("keygen-help-flag-output"))]
     output: Option<String>,
 }
 
@@ -57,40 +76,35 @@ fn main() -> Result<(), error::Error> {
     // Isolation Marks, so we disable them for now.
     LANGUAGE_LOADER.set_use_isolating(false);
 
-    let opts = AgeOptions::parse_args_default_or_exit();
+    let opts = AgeOptions::parse();
 
-    if opts.version {
-        println!("rage-keygen {}", env!("CARGO_PKG_VERSION"));
+    let mut output = file_io::OutputWriter::new(
+        opts.output,
+        false,
+        file_io::OutputFormat::Text,
+        0o600,
+        false,
+    )
+    .map_err(error::Error::FailedToOpenOutput)?;
+
+    let sk = age::x25519::Identity::generate();
+    let pk = sk.to_public();
+
+    (|| {
+        writeln!(
+            output,
+            "# {}: {}",
+            fl!("identity-file-created"),
+            chrono::Local::now().to_rfc3339_opts(chrono::SecondsFormat::Secs, true)
+        )?;
+        writeln!(output, "# {}: {}", fl!("identity-file-pubkey"), pk)?;
+        writeln!(output, "{}", sk.to_string().expose_secret())?;
+
+        if !output.is_terminal() {
+            eprintln!("{}: {}", fl!("tty-pubkey"), pk);
+        }
+
         Ok(())
-    } else {
-        let mut output = file_io::OutputWriter::new(
-            opts.output,
-            false,
-            file_io::OutputFormat::Text,
-            0o600,
-            false,
-        )
-        .map_err(error::Error::FailedToOpenOutput)?;
-
-        let sk = age::x25519::Identity::generate();
-        let pk = sk.to_public();
-
-        (|| {
-            writeln!(
-                output,
-                "# {}: {}",
-                fl!("identity-file-created"),
-                chrono::Local::now().to_rfc3339_opts(chrono::SecondsFormat::Secs, true)
-            )?;
-            writeln!(output, "# {}: {}", fl!("identity-file-pubkey"), pk)?;
-            writeln!(output, "{}", sk.to_string().expose_secret())?;
-
-            if !output.is_terminal() {
-                eprintln!("{}: {}", fl!("tty-pubkey"), pk);
-            }
-
-            Ok(())
-        })()
-        .map_err(error::Error::FailedToWriteOutput)
-    }
+    })()
+    .map_err(error::Error::FailedToWriteOutput)
 }